+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/composer/MessageComposer.kt
@@ -482,7 +482,7 @@ internal fun DefaultComposerLabel() {
  * @param messageComposerState The state of the message input.
  * @param onValueChange Handler when the input field value changes.
  * @param onAttachmentRemoved Handler when the user taps on the cancel/delete attachment action.
- * */
+ */
 @Composable
 public fun RowScope.DefaultComposerInputContent(
     messageComposerState: MessageComposerState,
@@ -519,7 +519,7 @@ internal fun DefaultMessageComposerTrailingContent(
     validationErrors: List<ValidationError>,
     onSendMessage: (String, List<Attachment>) -> Unit,
 ) {
-    val isInputValid = (value.isNotEmpty() || attachments.isNotEmpty()) && validationErrors.isEmpty()
+    val isInputValid = (value.isNotBlank() || attachments.isNotEmpty()) && validationErrors.isEmpty()
 
     if (coolDownTime > 0) {
         CoolDownIndicator(coolDownTime = coolDownTime)