+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/Debouncer.kt
@@ -21,7 +21,7 @@ public class Debouncer(private val debounceMs: Long) {
     /**
      * Cancels the previous work and launches a new coroutine
      * containing the new work.
-     * */
+     */
     public fun submit(work: () -> Unit) {
         job?.cancel()
         job = scope.launch {
@@ -33,7 +33,7 @@ public class Debouncer(private val debounceMs: Long) {
     /**
      * Cancels the previous work and launches a new coroutine
      * containing the new suspendable work.
-     * */
+     */
     public fun submitSuspendable(work: suspend () -> Unit) {
         job?.cancel()
         job = scope.launch {
@@ -44,7 +44,7 @@ public class Debouncer(private val debounceMs: Long) {
 
     /**
      * Cancels the current work without shutting down the Coroutine scope.
-     * */
+     */
     public fun cancelLastDebounce() {
         job?.cancel()
     }