+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/input/internal/MessageInputFieldView.kt
@@ -412,7 +412,7 @@ internal class MessageInputFieldView : FrameLayout {
         }
     }
 
-    private fun hasValidText(): Boolean = isMessageTextValid(messageText)
+    private fun hasValidText(): Boolean = MessageTextValidator.isMessageTextValid(messageText)
 
     fun hasValidContent(): Boolean {
         return hasValidText() || selectedAttachments.isNotEmpty() || selectedCustomAttachments.isNotEmpty()