+++ stream-chat-android-ui-components/src/test/kotlin/io/getstream/chat/android/ui/message/input/internal/MessageTextValidatorTest.kt
@@ -3,12 +3,12 @@ package io.getstream.chat.android.ui.message.input.internal
 import org.amshove.kluent.`should be`
 import org.junit.jupiter.api.Test
 
-internal class TextValidatorTest {
+internal class MessageTextValidatorTest {
 
     @Test
     fun `should not be possible to send empty giphys`() {
         val message = "/giphy"
-        val result = isMessageTextValid(message)
+        val result = MessageTextValidator.isMessageTextValid(message)
 
         result `should be` (false)
     }
@@ -16,7 +16,7 @@ internal class TextValidatorTest {
     @Test
     fun `should not be possible to send blank giphys`() {
         val message = "/giphy      "
-        val result = isMessageTextValid(message)
+        val result = MessageTextValidator.isMessageTextValid(message)
 
         result `should be` (false)
     }
@@ -24,7 +24,7 @@ internal class TextValidatorTest {
     @Test
     fun `should be possible to send giphys with content`() {
         val message = "/giphy hi"
-        val result = isMessageTextValid(message)
+        val result = MessageTextValidator.isMessageTextValid(message)
 
         result `should be` (true)
     }
@@ -32,7 +32,7 @@ internal class TextValidatorTest {
     @Test
     fun `should be possible to send text`() {
         val message = "hi"
-        val result = isMessageTextValid(message)
+        val result = MessageTextValidator.isMessageTextValid(message)
 
         result `should be` (true)
     }