+++ CHANGELOG.md
@@ -46,6 +46,8 @@
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
+- Users now able to open `MessageOptionsDialogFragment` by clicking on a reaction left on a Giphy message. [#3620](https://github.com/GetStream/stream-chat-android/pull/3260)
+- inside `MessageOptionsDialogFragment` now properly displays all of the reactions to a message. Previously it erroneously displayed a blank state. [#3620](https://github.com/GetStream/stream-chat-android/pull/3260)
 
 ### ⬆️ Improved