+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - inside `MessageOptionsDialogFragment` now properly displays all of the reactions to a message. Previously it erroneously displayed a blank state. [#3620](https://github.com/GetStream/stream-chat-android/pull/3260)
 - Fixed the links in UI Components code snippets. [#3261](https://github.com/GetStream/stream-chat-android/pull/3261)
 - Messages containing links are now properly aligned with other types of messages. They use `@dimen/stream_ui_spacing_small` for their root layout start and end padding. [#3264](https://github.com/GetStream/stream-chat-android/pull/3264)
+- Made it impossible to send blank or empty messages. [#3269](https://github.com/GetStream/stream-chat-android/pull/3269)
 
 ### ⬆️ Improved
 
@@ -62,6 +63,7 @@
 
 ## stream-chat-android-compose
 ### 🐞 Fixed
+- Made it impossible to send blank or empty messages. [#3269](https://github.com/GetStream/stream-chat-android/pull/3269)
 
 ### ⬆️ Improved